Update patch set 2
Patch Set 2:

The question is simply: who is in charge of calculating the results? I'm fine to let the testcase decide if really needed. But in that case it requires deeper changes and this class will simply list 3 dummy methods at the end. Then it may be simply removed.

Patch-set: 2