Remove unnecessary condition in statement flow
First, the condition should be limited to only defining the default
value, rather than applying to the statement flow. This not only avoids
unnecessary repetition of statements, but is more logical.

Second, it is not even necessary to explicitly code a condition here,
because the language itself defaults to our desired default value.

Signed-off-by: Vincent Danno <vincent.danno@orange.com>
Change-Id: I119567f5b31444a0e864ba41dee8a5f640ff0c22